feat(ideavim): add ideavim plugin for IntelliJ
Yeah, I'm sure about returning IntelliJ Community Edition IDE for my
work. The motivation here is very simple - I wanna write some Kotlin
code and the `kotlin-language-server`, unfortunately, very awful. But
the IntelliJ, even if it is Community Edition, surprised me with
rapidity and flawless work with Kotlin. So I decided move to it, but
need to put some vim mappings, to which I have habits.