feat!(nvim): general cleaning of config files
It's cursed that I'd the same functionality but with doubled LOC...
Now I'd ripped out a lot of useless code with replacing helpful neovim's
or plugin's functions that are more cleaner than my code mess.

Also improved some place of code which will be repsonsive than earlier.
Especially the rust filetype, some mappings of which-keys.